mahjong-vision-calculater/pyproject.toml
[tool.poetry]
name = "mahjong-vision-calculater"
version = "0.1.0"
description = ""
authors = ["JongunJeong <skeep194@knu.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
roboflow = "^1.1.7"
ultralytics = "^8.0.205"
pillow = "^10.1.0"
mahjong = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
